Wrap grant DDL in execWithRole and split customrole by domain
Extend the transactional SET LOCAL ROLE wrapper (previously only used for
function DDL) to SyncDatabaseGrants and RevokeAllDatabaseGrants, so a
failing middle statement can no longer return the connection to the pool
with an active SET ROLE.

Rename postgres.UntargetableDatabases to ReservedSystemDatabases and
clarify in its doc-comment why postgres is intentionally excluded.

Split pkg/postgres/customrole.go and its tests into roles/grants/functions
files so each logic file has a matching _test.go with the relevant tests,
and drop the empty internal/controller/customrole_*_test.go placeholders.
