--- internal/controller/customrole_functions_test.go
@@ -1,1 +1,0 @@
-package controller_test
