--- internal/controller/customrole_grants_test.go
@@ -1,1 +1,0 @@
-package controller_test
